requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
supabase>=2.0.0
greenlet>=2.0.0
email-validator>=2.0.0
